+++ CHANGELOG.md
@@ -6,6 +6,7 @@ language runtime. The main focus is on user-observable behavior of the engine.
 ## Version 21.1.0
 
 * Added subclassing of Java classes in JVM mode
+* Use native posix functions in the GraalPython Launcher
 
 ## Version 21.0.0